server/src/test/java/server/api/TestConnectionControllerTest.java
package server.api;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.MockitoAnnotations;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.ResponseEntity;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

@ExtendWith(MockitoExtension.class)
class TestConnectionControllerTest {

    @InjectMocks
    private TestConnectionController testConnectionController;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    void testConnection() {
        assertThat(testConnectionController.testConnection())
                .isEqualTo(ResponseEntity.ok("Successful"));
    }
}